NIFI-1728 Specifying charsets for messages sent to/received from Kafka in Kafka processor and related tests to remedy failures in Windows environments. Specifying EOF as an int instead of a byte.

+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-processors/src/main/java/org/apache/nifi/processors/kafka/StreamScanner.java
@@ -26,7 +26,7 @@ import java.nio.ByteBuffer;
  */
 class StreamScanner {
 
-    private final static byte EOF = -1;
+    private final static int EOF = -1;
 
     private final InputStream is;
